Create day 1-3
This repository is a comprehensive collection of my Java learning journey, focusing on Object-Oriented Programming (OOP) and core Java concepts. All the code, exercises, and mini-projects are designed to reinforce my understanding of Java fundamentals and good programming practices.

The topics in this repository are studied from “Head First Java” by Kathy Sierra & Bert Bates, which helped me understand how Java works, the anatomy of a class, and OOP principles in a hands-on way.